Add docblocks to order properties, move element specific methods to trait.
